localserve-ai/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
google-genai==1.4.0
pandas==2.2.3
numpy==2.1.2
python-dotenv==1.0.1
requests==2.32.3
